Correct Default SPL_SIZE_LIMIT for Cyclone/Arria V
Refer to the Cyclone V Hard Processor System Technical Reference Manual (same for Arria V), Section A-32, "Loading the Preloader Image".
The upper 4kB of the on-chip RAM is reserved by the Boot ROM, meaning only 60kB is available for the preloader. If the SPL extends into this region, it causes the boot ROM to crash.